r/bgp|rip: centralize resource schema of common block
in group and neighbor
