+++ internal/providerfwk/upgradestate_bgp_neighbor.go
@@ -219,12 +219,12 @@ func (rsc *bgpNeighbor) UpgradeState(_ context.Context) map[int64]resource.State
 							Blocks: map[string]schema.Block{
 								"accepted_prefix_limit": schema.ListNestedBlock{
 									NestedObject: schema.NestedBlockObject{
-										Attributes: rsc.schemaFamilyPrefixLimitAttributes(),
+										Attributes: bgpBlockFamilyBlockPrefixLimit{}.resourceSchemaAttributes(),
 									},
 								},
 								"prefix_limit": schema.ListNestedBlock{
 									NestedObject: schema.NestedBlockObject{
-										Attributes: rsc.schemaFamilyPrefixLimitAttributes(),
+										Attributes: bgpBlockFamilyBlockPrefixLimit{}.resourceSchemaAttributes(),
 									},
 								},
 							},
@@ -240,12 +240,12 @@ func (rsc *bgpNeighbor) UpgradeState(_ context.Context) map[int64]resource.State
 							Blocks: map[string]schema.Block{
 								"accepted_prefix_limit": schema.ListNestedBlock{
 									NestedObject: schema.NestedBlockObject{
-										Attributes: rsc.schemaFamilyPrefixLimitAttributes(),
+										Attributes: bgpBlockFamilyBlockPrefixLimit{}.resourceSchemaAttributes(),
 									},
 								},
 								"prefix_limit": schema.ListNestedBlock{
 									NestedObject: schema.NestedBlockObject{
-										Attributes: rsc.schemaFamilyPrefixLimitAttributes(),
+										Attributes: bgpBlockFamilyBlockPrefixLimit{}.resourceSchemaAttributes(),
 									},
 								},
 							},
@@ -261,12 +261,12 @@ func (rsc *bgpNeighbor) UpgradeState(_ context.Context) map[int64]resource.State
 							Blocks: map[string]schema.Block{
 								"accepted_prefix_limit": schema.ListNestedBlock{
 									NestedObject: schema.NestedBlockObject{
-										Attributes: rsc.schemaFamilyPrefixLimitAttributes(),
+										Attributes: bgpBlockFamilyBlockPrefixLimit{}.resourceSchemaAttributes(),
 									},
 								},
 								"prefix_limit": schema.ListNestedBlock{
 									NestedObject: schema.NestedBlockObject{
-										Attributes: rsc.schemaFamilyPrefixLimitAttributes(),
+										Attributes: bgpBlockFamilyBlockPrefixLimit{}.resourceSchemaAttributes(),
 									},
 								},
 							},
